+++ packages/core/CHANGELOG.md
@@ -1,3 +1,5 @@
 # @super-admin-org/core
 
+## 0.1.2
+
 ## 0.1.1