+++ packages/theme-cyberpunk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @super-admin-org/theme-cyberpunk
 
+## 0.1.2
+
+### Patch Changes
+
+- @super-admin-org/core@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes