+++ packages/ui/CHANGELOG.md
@@ -1,3 +1,5 @@
 # @super-admin-org/ui
 
+## 0.1.2
+
 ## 0.1.1